requirements.txt
fastapi
uvicorn
psycopg2-binary
pgvector
pypdf
langchain-google-genai
langchain-text-splitters
langchain-core
langgraph
python-dotenv
requests
